Move depviewer widget to main notebook extension
